requirements.txt
# Core dependencies
pandas
requests
openai
python-dotenv

# Supabase client
supabase

# Data processing
numpy
